requirements-dev.txt
black==23.11.0
click==8.1.7
factory-boy==3.3.0
Faker==20.0.0
flake8==6.1.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
six==1.16.0
